refactor: TripServiceForScheduler 분리 [AIR-134]
