+++ aircnc/src/main/java/com/gurudev/aircnc/domain/trip/service/TripService.java
@@ -14,8 +14,4 @@ public interface TripService {
 
   Trip cancel(Long tripId, Long guestId);
 
-  void bulkStatusToTravelling();
-
-  void bulkStatusToDone();
-
 }
