+++ aircnc/src/main/java/com/gurudev/aircnc/domain/trip/service/TripServiceForScheduler.java
@@ -1,0 +1,8 @@
+package com.gurudev.aircnc.domain.trip.service;
+
+public interface TripServiceForScheduler {
+
+  void bulkStatusToTravelling();
+
+  void bulkStatusToDone();
+}
